Type saleor/checkout for mypy --strict
Add full annotations across the checkout package (models, utils, fetch,
calculations, delivery_context, complete_checkout, tasks, webhooks,
search) and minimal signature annotations on the shared callees the
package uses (MoneyField/TaxedMoneyField, NotApplicable, PaymentError,
NoValidationSearchVector/FlatConcat, call_event, to_global_id_or_none,
LineInfo generics, etc.) so that the strict in-package error count
drops to zero.

Latent bugs exposed by typing and fixed here:

* complete_checkout_pre_payment_part: the `except Checkout.DoesNotExist`
  branch returned a bare Order instead of the declared 3-tuple, which
  would raise TypeError in the caller. The outer complete_checkout now
  catches the exception and returns `(order, False, {})`, matching the
  existing "checkout disappeared" handling further down.
* checkoutCreate / checkoutCreateFromOrder passed
  `site.settings.limit_quantity_per_checkout` positionally into
  `add_variants_to_checkout`, where it landed in the `replace: bool`
  parameter. The stray argument is removed so `replace` keeps its
  default.

+++ saleor/order/utils.py
@@ -450,7 +450,7 @@ def add_gift_cards_to_order(
     total_before_gift_card_compensation = total_price_left
     order_gift_cards = []
     gift_cards_to_update = []
-    balance_data: list[tuple[GiftCard, float]] = []
+    balance_data: list[tuple[GiftCard, Decimal]] = []
     used_by_user = checkout_info.user
     used_by_email = cast(str, checkout_info.get_customer_email())
     for gift_card in checkout_info.checkout.gift_cards.select_for_update():
@@ -498,7 +498,7 @@ def add_gift_cards_to_order(
 def update_gift_card_balance(
     gift_card: GiftCard,
     total_price_left: Money,
-    balance_data: list[tuple[GiftCard, float]],
+    balance_data: list[tuple[GiftCard, Decimal]],
 ) -> Money:
     previous_balance = gift_card.current_balance
     if total_price_left < gift_card.current_balance:
